Use `u32` for degree, fix `impl Arbitrary for Instance` (#170)
Split from #162 

- Use `u32` for polynomial degree
- Fix `impl Arbitrary for Instance`. `max_degree` and `max_id` are
reversed, bug of #163

+++ rust/ommx/src/convert/quadratic.rs
@@ -61,7 +61,7 @@ impl Quadratic {
         self.as_linear()?.as_constant()
     }
 
-    pub fn degree(&self) -> usize {
+    pub fn degree(&self) -> u32 {
         if self.values.iter().any(|v| v.abs() > f64::EPSILON) {
             2
         } else {
